+++ Changelog.md
@@ -19,3 +19,4 @@
 
 * `Missions:crewForRent` now requires the player to have `Player:withMenu` set
 * added `Player:withQuickDial()` to display quick dials on the relay station
+* `Station:withStorage` can now handle Scan Probes as products. Use the id `scanProbe`.